Moves single example to package
