+++ src/life/drewmiley/StreamExamplesRunner.java
@@ -1,8 +1,8 @@
 package life.drewmiley;
 
-import life.drewmiley.examples.*;
 import life.drewmiley.examples.helper.ExampleRunner;
 import life.drewmiley.examples.helper.SimpleObject;
+import life.drewmiley.examples.single.*;
 
 import java.util.List;
