+++ src/life/drewmiley/examples/single/AnyMatch.java
@@ -1,14 +1,14 @@
-package life.drewmiley.examples;
+package life.drewmiley.examples.single;
 
 import life.drewmiley.examples.helper.SimpleObject;
 
 import java.util.Arrays;
 import java.util.List;
 
-public class AllMatch {
+public class AnyMatch {
 
-    private boolean niceFilter(SimpleObject simpleObject) {
-        return simpleObject.getText() != null;
+    private boolean evilFilter(SimpleObject simpleObject) {
+        return simpleObject.getNumber() > 1000000;
     }
 
     private boolean arbitraryFilter(SimpleObject simpleObject) {
@@ -17,34 +17,34 @@ public class AllMatch {
 
     public boolean runStreamForArray(SimpleObject[] simpleObjects) {
         return Arrays.stream(simpleObjects)
-                .allMatch(simpleObject -> arbitraryFilter(simpleObject));
+                .anyMatch(simpleObject -> arbitraryFilter(simpleObject));
     }
 
     public boolean runImperativeForArray(SimpleObject[] simpleObjects) {
-        boolean allMatch = true;
+        boolean anyMatch = false;
         for (int i = 0; i < simpleObjects.length; i++) {
             SimpleObject simpleObject = simpleObjects[i];
-            if (!arbitraryFilter(simpleObject)) {
-                allMatch = false;
+            if (arbitraryFilter(simpleObject)) {
+                anyMatch = true;
                 break;
             }
         }
-        return allMatch;
+        return anyMatch;
     }
 
     public boolean runStreamForList(List<SimpleObject> simpleObjectList) {
         return simpleObjectList.stream()
-                .allMatch(simpleObject -> arbitraryFilter(simpleObject));
+                .anyMatch(simpleObject -> arbitraryFilter(simpleObject));
     }
 
     public boolean runImperativeForList(List<SimpleObject> simpleObjectList) {
-        boolean allMatch = true;
+        boolean anyMatch = false;
         for (SimpleObject simpleObject : simpleObjectList) {
-            if (!arbitraryFilter(simpleObject)) {
-                allMatch = false;
+            if (arbitraryFilter(simpleObject)) {
+                anyMatch = true;
                 break;
             }
         }
-        return allMatch;
+        return anyMatch;
     }
 }
